fetch failure integration testing (#46)
