+++ tests/spark-it/src/test/scala/org/apache/celeborn/tests/spark/SparkTestBase.scala
@@ -17,19 +17,26 @@
 
 package org.apache.celeborn.tests.spark
 
+import java.io.File
+import java.util.concurrent.atomic.AtomicBoolean
+
 import scala.util.Random
 
-import org.apache.spark.SPARK_VERSION
-import org.apache.spark.SparkConf
+import org.apache.spark.{SPARK_VERSION, SparkConf, TaskContext}
+import org.apache.spark.shuffle.ShuffleHandle
+import org.apache.spark.shuffle.celeborn.{CelebornShuffleHandle, ShuffleManagerHook, SparkUtils}
 import org.apache.spark.sql.SparkSession
 import org.apache.spark.sql.internal.SQLConf
 import org.scalatest.{BeforeAndAfterAll, BeforeAndAfterEach}
 import org.scalatest.funsuite.AnyFunSuite
 
+import org.apache.celeborn.client.ShuffleClient
+import org.apache.celeborn.common.CelebornConf
 import org.apache.celeborn.common.CelebornConf._
 import org.apache.celeborn.common.internal.Logging
 import org.apache.celeborn.common.protocol.ShuffleMode
 import org.apache.celeborn.service.deploy.MiniClusterFeature
+import org.apache.celeborn.service.deploy.worker.Worker
 
 trait SparkTestBase extends AnyFunSuite
   with Logging with MiniClusterFeature with BeforeAndAfterAll with BeforeAndAfterEach {
@@ -50,6 +57,16 @@ trait SparkTestBase extends AnyFunSuite
   override def afterAll(): Unit = {
     logInfo("all test complete , stop Celeborn mini cluster")
     shutdownMiniCluster()
+  }
+
+  var workerDirs: Seq[String] = Seq.empty
+
+  override def createWorker(map: Map[String, String]): Worker = {
+    val storageDir = createTmpDir()
+    this.synchronized {
+      workerDirs = workerDirs :+ storageDir
+    }
+    super.createWorker(map, storageDir)
   }
 
   def updateSparkConf(sparkConf: SparkConf, mode: ShuffleMode): SparkConf = {
@@ -97,5 +114,46 @@ trait SparkTestBase extends AnyFunSuite
     val result = sparkSession.sql("select fa,count(fb) from tmp group by fa order by fa")
     val outMap = result.collect().map(row => row.getString(0) -> row.getLong(1)).toMap
     outMap
+  }
+
+  class ShuffleReaderFetchFailureGetHook(conf: CelebornConf) extends ShuffleManagerHook {
+    var executed: AtomicBoolean = new AtomicBoolean(false)
+    val lock = new Object
+
+    override def exec(
+        handle: ShuffleHandle,
+        startPartition: Int,
+        endPartition: Int,
+        context: TaskContext): Unit = {
+      if (executed.get() == true) return
+
+      lock.synchronized {
+        handle match {
+          case h: CelebornShuffleHandle[_, _, _] => {
+            val appUniqueId = h.appUniqueId
+            val shuffleClient = ShuffleClient.get(
+              h.appUniqueId,
+              h.lifecycleManagerHost,
+              h.lifecycleManagerPort,
+              conf,
+              h.userIdentifier,
+              h.extension)
+            val celebornShuffleId = SparkUtils.celebornShuffleId(shuffleClient, h, context, false)
+            val allFiles = workerDirs.map(dir => {
+              new File(s"$dir/celeborn-worker/shuffle_data/$appUniqueId/$celebornShuffleId")
+            })
+            val datafile = allFiles.filter(_.exists())
+              .flatMap(_.listFiles().iterator).sortBy(_.getName).headOption
+            datafile match {
+              case Some(file) => file.delete()
+              case None => throw new RuntimeException("unexpected, there must be some data file" +
+                  s" under ${workerDirs.mkString(",")}")
+            }
+          }
+          case _ => throw new RuntimeException("unexpected, only support RssShuffleHandle here")
+        }
+        executed.set(true)
+      }
+    }
   }
 }
